Override getMessage in UnexpectedStatus
Before this change, standard approaches to logging errors would result
in something that looked like `org.http4s.client.UnexpectedStatus:
null`, which is not the most helpful error message in the world. I don't
have strong opinions on what this message should look like and I'm happy
to change it, but I think that it should include the status (at least
its code) in some form.