Remove `Content-Length` and `Content-Encoding` when decompressing
When the gzip middleware decompresses a response, these headers are no
longer correct, so it seems best to remove them. This is crucial when
the client is used as a proxy (or, in my case, as an STTP backend) to
prevent downstream consumers attempting to decompress the response
again.

This change was previously introduced here:
https://github.com/http4s/http4s/pull/2673http4s/http4s#discussion_r298684594 but was
removed (possibly unintentionally?) in this PR: https://github.com/http4s/http4s/pull/3476